Solved a bunch of tasks. Added utilities for numbers/math/combinatorics

+++ algo-commons/src/main/java/com/rqpa/algo/primes/PrimeNumberChecker.java
@@ -2,5 +2,5 @@ package com.rqpa.algo.primes;
 
 public interface PrimeNumberChecker
 {
-    boolean isPrime(int n);
+    boolean isPrime(long n);
 }
